+++ sdk/formrecognizer/Azure.AI.FormRecognizer/tests/samples/Sample6_ManageCustomModels.cs
@@ -49,10 +49,10 @@ namespace Azure.AI.FormRecognizer.Samples
 
             Console.WriteLine($"Custom Model {modelCopy.ModelId} recognizes the following form types:");
 
-            foreach (CustomFormSubModel subModel in modelCopy.Submodels)
+            foreach (CustomFormSubmodel submodel in modelCopy.Submodels)
             {
-                Console.WriteLine($"SubModel Form Type: {subModel.FormType}");
-                foreach (CustomFormModelField field in subModel.Fields.Values)
+                Console.WriteLine($"Submodel Form Type: {submodel.FormType}");
+                foreach (CustomFormModelField field in submodel.Fields.Values)
                 {
                     Console.Write($"    FieldName: {field.Name}");
                     if (field.Label != null)