+++ sdk/formrecognizer/Azure.AI.FormRecognizer/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Property `CreatedOn` in types `CustomFormModel` and `CustomFormModelInfo` renamed to `RequestedOn`.
 - Property `LastModified` in types `CustomFormModel` and `CustomFormModelInfo` renamed to `CompletedOn`.
 - Property `Models` in `CustomFormModel` renamed to `Submodels`.
+- Type `CustomFormSubModel` renamed to `CustomFormSubmodel`.
 - `ContentType` renamed to `FormContentType`.
 - Parameter `useLabels` in `FormTrainingClient.StartTraining` renamed to `useTrainingLabels`.
 - Parameter `trainingFiles` in `FormTrainingClient.StartTraining` renamed to `trainingFilesUri`.